requirements.txt
python-telegram-bot
pandas-datareader
fix-yahoo-finance